scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
# Optional: for sources that block simple HTTP (WWR, Dynamite Jobs, Jobspresso)
playwright>=1.40.0
# Load .env for REMOTESOURCE_* credentials
python-dotenv>=1.0.0
